+++ buckifier/buckify_rocksdb.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
 from __future__ import absolute_import
 from __future__ import division
@@ -184,6 +185,10 @@ def generate_targets(repo_path, deps_map):
         src_mk.get("ANALYZER_LIB_SOURCES", [])
         + src_mk.get('STRESS_LIB_SOURCES', [])
         + ["test_util/testutil.cc"])
+    # db_stress binary
+    TARGETS.add_binary("db_stress",
+                       ["db_stress_tool/db_stress.cc"],
+                       [":rocksdb_stress_lib"])
     # bench binaries
     for src in src_mk.get("MICROBENCH_SOURCES", []):
         name =  src.rsplit('/',1)[1].split('.')[0] if '/' in src else src.split('.')[0]